Crate polkavm_common

This is an internal crate of PolkaVM; do not use it directly unless you absolutely know what you’re doing.

Modules

abi

Everything in this module affects the ABI of the guest programs, either by affecting their observable behavior (no matter how obscure), or changing which programs are accepted by the VM.

assembler

cast

This module defines an explicit casting facility to replace Rust’s built-in as casts. The general idea is:

elf

hasher

operation

program

regmap

utils

varint

writer

zygote

This module defines the ABI boundary between the host and the zygote.

Macros

build_static_dispatch_table

static_assert