Module polkavm_common::zygote

This module defines the ABI boundary between the host and the zygote.

In general everything here can be modified at will, provided the zygote is recompiled.

Structs

AddressTable

AddressTablePacked

AddressTableRaw

CacheAligned

ExtTable

ExtTablePacked

ExtTableRaw

VmCtx

The virtual machine context.

VmCtxCounters

VmCtxHeapInfo

VmInit

VmMap

Constants

VMCTX_FUTEX_BUSY

The VM is busy.

VMCTX_FUTEX_GUEST_ECALLI

The VM has triggered a host call and is idle.

VMCTX_FUTEX_GUEST_SIGNAL

The VM’s signal handler was triggered.

VMCTX_FUTEX_GUEST_STEP

The VM has went through a single instruction is idle.

VMCTX_FUTEX_GUEST_TRAP

The VM has triggered a trap and is idle.

VMCTX_FUTEX_IDLE

The VM is idle.

VM_ADDR_JUMP_TABLE

The address where the indirect jump table starts inside of the VM.

VM_ADDR_JUMP_TABLE_RETURN_TO_HOST

The address where the return-to-host jump table vector physically resides.

VM_ADDR_NATIVE_CODE

The address where the native code starts inside of the VM.

VM_ADDR_NATIVE_STACK_HIGH

The address of the top of the native stack.

VM_ADDR_NATIVE_STACK_LOW

The address of the native stack.

VM_ADDR_NATIVE_STACK_SIZE

The size of the native stack.

VM_ADDR_SHARED_MEMORY

Address where the shared memory is mapped.

VM_ADDR_SIGSTACK

The address of the signal stack.

VM_ADDR_VMCTX

The address of the global per-VM context struct.

VM_COMPILER_MAXIMUM_INSTRUCTION_LENGTH

The maximum number of native code bytes that can be emitted by a single VM instruction.

VM_SANDBOX_MAXIMUM_JUMP_TABLE_SIZE

The maximum number of bytes the jump table can be.

VM_SANDBOX_MAXIMUM_JUMP_TABLE_VIRTUAL_SIZE

The maximum number of bytes the jump table can span in virtual memory.

VM_SANDBOX_MAXIMUM_NATIVE_CODE_SIZE

The maximum number of bytes the native code can be.

VM_SHARED_MEMORY_SIZE

The size of the shared memory region.