Crate polkavm_common

This is an internal crate of PolkaVM; do not use it directly unless you absolutely know what you’re doing.

Modules

abi

Everything in this module affects the ABI of the guest programs, either by affecting their observable behavior (no matter how obscure), or changing which programs are accepted by the VM.

elf

error

operation

program

utils

varint

writer

zygote

This module defines the ABI boundary between the host and the zygote.

Macros

implement_instruction_visitor

prepare_visitor

static_assert

Constants

HOSTCALL_TRACE

A special hostcall number set by the guest to trigger a trace.

VM_RPC_FLAG_CLEAR_PROGRAM_AFTER_EXECUTION

A flag which will trigger the sandbox to unload its program after execution.

VM_RPC_FLAG_RESET_MEMORY_AFTER_EXECUTION

A flag which will trigger the sandbox to reset its memory after execution.

VM_RPC_FLAG_RESET_MEMORY_BEFORE_EXECUTION

A flag which will trigger the sandbox to reset its memory before execution.