Crate polkavm_common

This is an internal crate of PolkaVM; do not use it directly unless you absolutely know what you’re doing.

Modules

abi

Everything in this module affects the ABI of the guest programs, either by affecting their observable behavior (no matter how obscure), or changing which programs are accepted by the VM.

assembler

cast

This module defines an explicit casting facility to replace Rust’s built-in as casts. The general idea is:

elf

hasher

operation

program

regmap

simulator

utils

varint

writer

zygote

This module defines the ABI boundary between the host and the zygote.

Macros

build_static_dispatch_table_jam_v1

build_static_dispatch_table_latest32

build_static_dispatch_table_latest64

build_static_dispatch_table_revive_v1

impl_parsing_visitor_for_instruction_visitor

static_assert