polkavm_common

Module abi

Everything in this module affects the ABI of the guest programs, either by affecting their observable behavior (no matter how obscure), or changing which programs are accepted by the VM.

Structs

MemoryMap

The memory map of a given guest program.

Constants

VM_ADDR_RETURN_TO_HOST

The address which, when jumped to, will return to the host.

VM_ADDR_USER_STACK_HIGH

The address at which the program’s stack starts inside of the VM.

VM_CODE_ADDRESS_ALIGNMENT

The minimum required alignment of runtime code pointers.

VM_MAXIMUM_EXPORT_COUNT

The maximum number of functions the program can export.

VM_MAXIMUM_IMPORT_COUNT

The maximum number of functions the program can import.

VM_MAXIMUM_INSTRUCTION_COUNT

The maximum number of VM instructions a program can be composed of.

VM_MAX_PAGE_SIZE

The maximum page size of the VM.

VM_MIN_PAGE_SIZE

The minimum page size of the VM.