polkavm_common

Module abi

Everything in this module affects the ABI of the guest programs, either by affecting their observable behavior (no matter how obscure), or changing which programs are accepted by the VM.

Structs

MemoryMap

The memory map of a given guest program.

MemoryMapBuilder

Constants

VM_ADDR_RETURN_TO_HOST

The address which, when jumped to, will return to the host.

VM_CODE_ADDRESS_ALIGNMENT

The minimum required alignment of runtime code pointers.

VM_MAXIMUM_CODE_SIZE

The maximum byte size of the code blob.

VM_MAXIMUM_IMPORT_COUNT

The maximum number of functions the program can import.

VM_MAXIMUM_JUMP_TABLE_ENTRIES

The maximum number of entries in the jump table.

VM_MAX_PAGE_SIZE

The maximum page size of the VM.

VM_MIN_PAGE_SIZE

The minimum page size of the VM.