polkavm_common

Module zygote

This module defines the ABI boundary between the host and the zygote.

In general everything here can be modified at will, provided the zygote is recompiled.

Structs

AddressTable

AddressTablePacked

AddressTableRaw

CacheAligned

SandboxMemoryConfig

The memory configuration used by a given program and/or sandbox instance.

VmCtx

The virtual machine context.

VmCtxCounters

VmCtxHeapInfo

VmCtxSyscall

VmInit

Constants

HOSTCALL_ABORT_EXECUTION

A special hostcall number set by the host to signal that the guest should stop executing the program.

HOSTCALL_SBRK

A special hostcall number set by the host to signal that the guest should execute sbrk.

SANDBOX_EMPTY_NATIVE_PROGRAM_COUNTER

A sentinel value to indicate that the native program counter is not available.

SANDBOX_EMPTY_NTH_INSTRUCTION

A sentinel value to indicate that the instruction counter is not available.

VMCTX_FUTEX_BUSY

The VM is busy.

VMCTX_FUTEX_HOSTCALL

The VM has triggered a host call.

VMCTX_FUTEX_IDLE

The VM is idle and is waiting for work.

VMCTX_FUTEX_INIT

The VM is ready to be initialized.

VMCTX_FUTEX_TRAP

The VM has triggered a trap.

VM_ADDR_JUMP_TABLE

The address where the indirect jump table starts inside of the VM.

VM_ADDR_JUMP_TABLE_RETURN_TO_HOST

The address where the return-to-host jump table vector physically resides.

VM_ADDR_NATIVE_CODE

The address where the native code starts inside of the VM.

VM_ADDR_NATIVE_STACK_HIGH

The address of the top of the native stack.

VM_ADDR_NATIVE_STACK_LOW

The address of the native stack.

VM_ADDR_NATIVE_STACK_SIZE

The size of the native stack.

VM_ADDR_SIGSTACK

The address of the signal stack.

VM_ADDR_VMCTX

The address of the global per-VM context struct.

VM_COMPILER_MAXIMUM_EPILOGUE_LENGTH

The maximum number of native code bytes that can be emitted as an epilogue.

VM_COMPILER_MAXIMUM_INSTRUCTION_LENGTH

The maximum number of native code bytes that can be emitted by a single VM instruction.

VM_RPC_FLAG_RECONFIGURE

A flag which will trigger the sandbox to reload its program before execution.

VM_SANDBOX_MAXIMUM_JUMP_TABLE_SIZE

The maximum number of bytes the jump table can be.

VM_SANDBOX_MAXIMUM_JUMP_TABLE_VIRTUAL_SIZE

The maximum number of bytes the jump table can span in virtual memory.

VM_SANDBOX_MAXIMUM_NATIVE_CODE_SIZE

The maximum number of bytes the native code can be.